More tests for emap; hey, we found another bug!
When the arguments to an expression are more expressions,
we make a recursive call to emap to process them.  That if they're not?
You were returning them unchanged; that's wrong.  The right thing
to do is to ignore them, because the callback you will invoke on the current
expression will handle them if necessary.