Minor tinkering with labelbox definitions in view of recent change to grammar---param-specs are now allowed to be compound names, so we no longer need weird legerdemain to make sure that you can set the label in a labelbox; just doing
	labelbox L(the_label.label = "foo")

will work.  Except this is nasty-looking, so I renamed some things so
that it's now

	labelbox L(label.text = "foo")

When labels get more complicated, we will also be able to do

	labelbox L(label.text = "foo",
		   label.color = "red")

etc.  I left in "text" as an abbreviation for "label.text".

I had to rewrite the labelbox-001 test to match this new style;
labelbox-002 tests to make sure the old definition would still have worked.