Begin to fix handling of compound names
There is a new, nearly-trivial class "Chunk", which encapsulates a
named type.  Subfeatures of a type are now "Chunks".  They are stored
in the {O} member as before, but indexed by the stringized version of
the name portion of the chunk.

Tests still don't pass, because there is confusion in the code between
arrays and non-arrays, but it's getting closer.