my_subchunks now handles array chunks
This necessitated adding an environment parameter to my_subchunks,
since it is needed to calculate the array bounds.
all_leaf_subchunks got a similar parameter.

XXX: Code for generating array element names is now duplicated in
my_subchunks and also in Type::Array; maybe this is unimportant, since
the former code is (I think) used only to generate variable names for
solving.

Cleaned up discrepancy in calling conventions of various *::bounds
methods.

The "count" argument in ->add_subchunk($name, $type, $count) should be
an expression, but some code was assuming a number.  Now if it is a
number, that is syntactic sugar for a constant expression.